Don't delete hidden groups during file cleanup

+++ CarcassSpark/ObjectViewers/ModViewerTabControl.cs
@@ -619,9 +619,19 @@ namespace CarcassSpark.ObjectViewers
 
         private void ClearContentFolder(string modLocation)
         {
+            Dictionary<string, string[]> hiddenGroups = Content.CustomManifest["hiddenGroups"]?.ToObject<Dictionary<string, string[]>>();
+            List<string> allHiddenGroups = new List<string>();
+            foreach(string key in hiddenGroups.Keys)
+            {
+                allHiddenGroups.AddRange(hiddenGroups[key]);
+            }
+
             foreach (string file in Directory.EnumerateFiles(modLocation + "/content/", "*.json"))
             {
-                File.Delete(file);
+                if (!allHiddenGroups.Contains(file))
+                {
+                    File.Delete(file);
+                }
             }
         }
 
@@ -633,6 +643,7 @@ namespace CarcassSpark.ObjectViewers
         public void SaveMod(string location)
         {
             CreateDirectories(location);
+            // TODO: prevent people from making groups that already exist, but are hidden
             ClearContentFolder(location);
             if (Content.Aspects.Count > 0)
             {
